project/requirements.txt
# Web 框架
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

# HTTP 客户端
httpx==0.27.2
aiohttp==3.10.10

# 论文搜索
arxiv==2.1.3
biopython==1.84          # PubMed 搜索
semanticscholar==0.8.4   # Semantic Scholar

# AI / LLM
openai==1.51.0           # DeepSeek 使用 OpenAI 兼容接口

# 向量存储 & Embedding
chromadb==0.5.23
sentence-transformers==3.2.1  # 本地 Embedding 模型
torch==2.4.1             # sentence-transformers 依赖（CPU 版，生产环境按需替换）

# 工具库
python-dotenv==1.0.1
pyyaml==6.0.2
tenacity==9.0.0          # 重试机制
loguru==0.7.2            # 结构化日志
numpy==1.26.4

# 邮件推送
aiosmtplib==3.0.2

# 测试
pytest==8.3.3
pytest-asyncio==0.24.0
